added everyone's changes into my project

+++ FeelingGoodApp/FeelingGoodApp/Services/Models/ExerciseInfo.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public string user_input { get; set; }
-        public float duration_min { get; set; }
-        public float met { get; set; }
-        public float nf_calories { get; set; }
+        public int duration_min { get; set; }
+        public int met { get; set; }
+        public int nf_calories { get; set; }
         //public ExercisePhoto photo { get; set; }
         //public int compendium_code { get; set; }
         public string name { get; set; }
